+++ doc/whats_new.rst
@@ -27,6 +27,7 @@ Enhancements
   command is available after executing ``%load_ext xsimlab.ipython``.
 - Added an optional cache for on-demand variables (:issue:`156`). The ``@compute``
   decorator now has a ``cache`` option (deactivated by default).
+- Added :func:`~xsimlab.group_dict` variable (:issue:`159`).
 
 Bug fixes
 ~~~~~~~~~